add a sample data for testing and reorganize generate codes
