+++ train_text2text.sh
@@ -7,15 +7,15 @@ python3 text2text/train.py \
     --train_file data/trec-pds.train.product2query.jsonl \
     --max_src_length 256  \
     --max_tgt_length 32 \
-    --output_dir /models/t5-base-product2query \
+    --output_dir models/t5-base-product2query \
     --do_train --do_eval \
-    --save_strategy steps \
     --max_steps 10000 \
     --save_steps 2500 \
     --eval_steps 500 \
+    --save_strategy steps \
     --evaluation_strategy steps \
-    --per_device_train_batch_size 32 \
-    --per_device_eval_batch_size 32 \
+    --per_device_train_batch_size 2 \
+    --per_device_eval_batch_size 2 \
     --optim adafactor \
     --learning_rate 1e-4 \
     --lr_scheduler_type constant \
